Revert "Fix for GCC sign-compare error for AzCore, AzFramework ( [-Werror=sign-compare] )"
This reverts commit d923f01382de29f9293804c0f2369348cab6e464.

+++ Code/Framework/AzCore/AzCore/Settings/CommandLine.cpp
@@ -169,7 +169,7 @@ namespace AZ
 
         // This version of Parse does not skip over 0th index
         AZStd::string currentSwitch;
-        for (size_t i = 0; i < commandLine.size(); ++i)
+        for (int i = 0; i < commandLine.size(); ++i)
         {
             AddArgument(commandLine[i], currentSwitch);
         }
